Move grpc mod into folder
