+++ tuxtape-server/src/tuxtape_database_bridge/grpc/mod.rs
@@ -1,0 +1,2 @@
+mod handlers;
+pub mod service;
